fake data was being written to the binary file. if a link element is nominally closed, even if there's actually flow calculated there, the flow is disregarded upon writing results. so when the flow is read back in, it's zero -- which is different than it would have been if it had stayed in memory. i wonder what that does to mass balance?
this version works, at least with Net3, simplenet, and SampleTown -- but the results are non-canonical, as the current workaround for the issue above writes tiny flow values to the binary file for closed elements. presumably we should check for closed status in the transport() function.

+++ src/types.h
@@ -28,7 +28,7 @@ AUTHOR:     L. Rossman
    Definition of 4-byte integers & reals
 -------------------------------------------
 */
-typedef  float        REAL4;                                                   //(2.00.11 - LR)
+typedef  double        REAL4;                                                   //(2.00.11 - LR)
 typedef  int          INT4;                                                    //(2.00.12 - LR)
 
 /*
